api/requirements.txt
bcrypt
fastapi[standard]
python-jose[cryptography] 
python-multipart
redis